+++ implants/lib/eldritchv2/eldritchv2/Cargo.toml
@@ -55,6 +55,7 @@ stdlib = [
 
 [dev-dependencies]
 httptest = "0.16"
+eldritchv2 = { workspace = true }
 
 [build-dependencies]
 regex = "1.10"
